requirements.txt
requests>=2.31.0
urllib3>=2.0.0
aiohttp>=3.9.0
docker>=6.1.0
PyYAML>=6.0.0
colorama>=0.4.6
